+++ src/main/java/by/itacademy/mikhalevich/universe/model/qualifiers/CustomStringQualifier.java
@@ -1,0 +1,15 @@
+package by.itacademy.mikhalevich.universe.model.qualifiers;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Qualifier
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.FIELD, ElementType.METHOD, ElementType.CONSTRUCTOR, ElementType.PARAMETER})
+public @interface CustomStringQualifier {
+    String name();
+}
